fix: optimistic UI + minor refactoring
